refactor: Answer 서비스에 대해서 CQS(Command Query Separation) 적용
