+++ src/main/java/com/codelion/animalcare/domain/doctorqna/controller/AnswerController.java
@@ -3,6 +3,7 @@ package com.codelion.animalcare.domain.doctorqna.controller;
 
 import com.codelion.animalcare.domain.doctorqna.dto.request.AnswerSaveRequestDto;
 import com.codelion.animalcare.domain.doctorqna.dto.request.AnswerUpdateRequestDto;
+import com.codelion.animalcare.domain.doctorqna.service.AnswerQueryService;
 import com.codelion.animalcare.domain.doctorqna.service.AnswerService;
 import com.codelion.animalcare.domain.doctorqna.service.QuestionService;
 import lombok.RequiredArgsConstructor;
@@ -23,6 +24,7 @@ import java.security.Principal;
 public class AnswerController {
 
     private final AnswerService answerService;
+    private final AnswerQueryService answerQueryService;
     private final QuestionService questionService;
 
     //답변 작성
@@ -45,7 +47,7 @@ public class AnswerController {
     @GetMapping("/usr/doctor-qna/{questionId}/answers/{answerId}/modify")
     public String modify(Model model, @PathVariable Long questionId, @PathVariable Long answerId, AnswerUpdateRequestDto answerUpdateRequestDto, Principal principal){
 
-        if(answerService.answerAuthorized(answerId, principal)){
+        if(answerQueryService.answerAuthorized(answerId, principal)){
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
         }
 
@@ -60,7 +62,7 @@ public class AnswerController {
             return "doctorqna/doctorQnaAnswerModifyForm";
         }
 
-        if(answerService.answerAuthorized(answerId, principal)){
+        if(answerQueryService.answerAuthorized(answerId, principal)){
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
         }
 
@@ -72,7 +74,7 @@ public class AnswerController {
     @GetMapping("/usr/doctor-qna/{questionId}/answers/{answerId}/delete")
     public String delete(@PathVariable Long questionId, @PathVariable Long answerId, Principal principal) {
 
-        if(answerService.answerAuthorized(answerId, principal)){
+        if(answerQueryService.answerAuthorized(answerId, principal)){
             throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "수정권한이 없습니다.");
         }
